back-end/requirements.txt
django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
pymongo==4.6.1
mongoengine==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
drf-yasg==1.21.7
django-storages==1.14.2
Pillow==10.1.0
python-magic==0.4.27